pyproject.toml
[project]
name = "secup-demo"
version = "0.1.0"
description = "SecUp - Traffic infraction management API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
]

[project.scripts]
start = "app.main:run"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
